server-python/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database - PostgreSQL
sqlalchemy==2.0.35
asyncpg==0.30.0
greenlet>=2.0.2

# Redis
redis[hiredis]==5.2.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# HTTP client
httpx==0.27.2

# LangChain
langchain==0.3.14
langchain-openai==0.3.0
langchain-core>=0.3.29,<0.4.0

# Data processing
numpy>=1.26.0
jieba==0.42.1
scikit-learn>=1.5.0
torch>=2.5.0

# pgvector
pgvector==0.3.6

# Scheduler
apscheduler==3.10.4

# Image processing
Pillow>=10.4.0

# Environment
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2
